effort-make-the-build-compile-against-the-vendo: burn-down round 1 - fix Point/Vector2 type mismatches in assigned files

+++ src/Murder/Core/Particles/Data/Emitter.cs
@@ -53,11 +53,11 @@ namespace Murder.Core.Particles
             }
             else if (Shape.Kind == EmitterShapeKind.Circle)
             {
-                return Rectangle.CenterRectangle(new Point(Shape.Circle.X, Shape.Circle.Y), Shape.Circle.Radius, Shape.Circle.Radius);
+                return Rectangle.CenterRectangle(new Point(Shape.Circle.X, Shape.Circle.Y), (int)Shape.Circle.Radius, (int)Shape.Circle.Radius);
             }
             else //if (Shape.Kind == EmitterShapeKind.CircleOutline)
             {
-                return Rectangle.CenterRectangle(new Point(Shape.Circle.X, Shape.Circle.Y), Shape.Circle.Radius, Shape.Circle.Radius);
+                return Rectangle.CenterRectangle(new Point(Shape.Circle.X, Shape.Circle.Y), (int)Shape.Circle.Radius, (int)Shape.Circle.Radius);
             }
         }
     }
